Move unused logic to dc_crate
